+++ src/nkdsl/__init__.py
@@ -55,9 +55,10 @@ from nkdsl.dsl import shift
 from nkdsl.dsl import shift_mod
 from nkdsl.dsl import site
 from nkdsl.dsl import emitted
-from nkdsl.dsl import global_index
+from nkdsl.dsl import source_index
 from nkdsl.dsl import swap
 from nkdsl.dsl import symbol
+from nkdsl.dsl import target_index
 from nkdsl.dsl import write
 
 #
@@ -146,7 +147,8 @@ __all__ = [
     "site",
     "emitted",
     "symbol",
-    "global_index",
+    "source_index",
+    "target_index",
     "SiteSelector",
     "AbstractIteratorClause",
     "AbstractPredicateClause",